requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
tensorflow==2.17.0
numpy==1.26.4
pillow==10.4.0
python-multipart==0.0.9